feat: 관리자 대시보드 컴포넌트 구조 개선
- 상품 관리 기능을 ProductManagement 컴포넌트로 분리
- AdminPage에 관리자 기능 통합 및 탭 관리 구현
- App.tsx 코드 간소화 (473줄 → 305줄)
