+++ src/basic/pages/Main/ShopPage/ShopPage.tsx
@@ -1,0 +1,3 @@
+export default function ShopPage() {
+  return <div>ShopPage</div>;
+}
